working prototype (requires compile goal)
