+++ asciidoctorj-extensions/src/main/java/Controller/Parameters.java
@@ -1,0 +1,26 @@
+/*
+ * To change this license header, choose License Headers in Project Properties.
+ * To change this template file, choose Tools | Templates
+ * and open the template in the editor.
+ */
+package Controller;
+
+/**
+ *
+ * @author LEVALLOIS
+ */
+public class Parameters {
+    
+    
+    public static String localDirectory = "C:\\Users\\levallois\\Google Drive\\open\\gephi tutorial\\";
+    public static String endOfPath = "history-of-gephi\\en\\";
+
+    public static String docName = "slide";
+    
+    public static boolean applyDefaultPicWidth = true;
+    public static boolean applyDefaultPicHeight = true;
+    
+    public static String defaultPicWidth = "300";
+    public static String defaultPicHeight = "300";
+    
+}
