+++ asciidoctorj-extensions/src/main/resources/META-INF/services/org.asciidoctor.extension.spi.ExtensionRegistry
