+++ docs/src/main/asciidoc/images/sunset.jpg
